paddle_service/requirements.txt
# PaddleOCR Mikroservis - Minimal Bağımlılıklar
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Image Processing (önce yükle - dependency order önemli)
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.6.0.66

# PaddleOCR ve bağımlılıkları
paddlepaddle==2.6.0
paddleocr==2.6.1.3

# Utilities
aiofiles==23.2.1
